Add user-defined chromosome order input
Previously, running find_translocations.py without a BED file led to an error
due to non-canonical chromosomes not being found in the list of chromosomes.
Now, the user provides a chromosome list which is interpreted. If the list
contains all chromosomes in the BED file, this fixes the error.

Also added sanity check that the number of reads parsed is the same as the
number of reads supposed to be in the pileup.